requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
google-cloud-storage
joblib
pillow